Upscale icons for better HiDPI support.
Custom icons were upscaled by re-exporting source svgs at higher resolution.

Fugue icons upscaled to 32x32 by simple pixel doubling. This way we are not
relying on OS scaling or on Qt to scale the icons, which may scale with
interpolation that looks even worse considering icons are pixel-art.

The script to perform the upscaling is included in the repository, and
the original 16x16 icons kept.

This is a decent stop-gap to look OK on high iDPI screens until an icon set as
comprehensive as fugue exists at higher resolution (I cannot find anything
comparable at the moment).

Including icons in multiple sizes was an option I explored, but it does not work
well with Qt designer - it does not choose the correct icon size from .ico
files. Including an excessively large icon and letting Qt downscale is the
current best practice (unfortunately). I would have liked to upscale the icons
using clever pixel-art scaling filters, but then they look terrible when
Qt *downscales* them fot a standard DPI screen. So scaling by pixel doubling
is the lesser evil for the moment.